+++ unstructured/ingest/connector/biomed.py
@@ -50,6 +50,7 @@ class SimpleBiomedConfig(BaseConnectorConfig):
     preserve_downloads: bool = False
     metadata_include: Optional[str] = None
     metadata_exclude: Optional[str] = None
+    fields_include: str = "element_id,text,type,metadata"
     flatten_metadata: bool = False
 
     def _validate_date_args(self, date):