+++ CHANGELOG.md
@@ -1,10 +1,13 @@
-## 0.5.7-dev0
+## 0.5.7-dev2
 
 ### Enhancements
+
+* Refactored codebase using `exactly_one`
 
 ### Features
 
 * Add `--flatten-metadata` parameter to `unstructured-ingest`
+* Add `--fields-include` parameter to `unstructured-ingest`
 
 ### Fixes